Clean Up Mac/Windows configs
Everything is set for my Mac now.

Had to change some of the Obsidian stuff. Splash screen opens a search
window to find a note instead of opening the Daily Note. I'll need a
Template to have it automatically open the daily note and right now the
NeoVim Obsidian can't handle my Obsidian Daily Note template that uses
templater. Maybe in the future.