refactor: update which-key bindings to new syntax
## Which-Key Syntax Update

They changed the syntax for which-key keybindings with v3. The old ones
would work for a while, but support will end eventually. All the
keybindings are updated to the new syntax.

## Setup Section

The `local setup` variable wasn't being used and the new syntax has
depricated a lot of those settings anyway. Tossed this section. Will use
the new setup if anything beyond the defaults are needed.